+++ src/features/auth/hooks/useLogin.ts
@@ -5,7 +5,7 @@ import { useNavigate } from 'react-router-dom';
 import { signInWithEmailAndPassword } from 'firebase/auth';
 import { auth } from '@/firebase';
 
-import { handleFirebaseError } from '@/utils/firebaseErrorHandler';
+import { handleFirebaseError } from '@/shared/utils/firebaseErrorHandler';
 
 interface LoginType {
   email: string;
