+++ src/shared/utils/firebaseErrorHandler.ts
